rename NavLInks -> NavLinks
